% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca_mean_distance_calculation.R
\name{analysis_pca_mean_distance_calculation}
\alias{analysis_pca_mean_distance_calculation}
\title{Given a set of coordinates, calculate the mean distance between all points.}
\usage{
analysis_pca_mean_distance_calculation(MATRIX)
}
\arguments{
\item{MATRIX}{- matrix with the rows refering to the points with N columns containing the coordinates (and therefore N dimensions).}
}
\value{
The mean distance (\code{numeric}).
}
\description{
Given a set of coordinates, calculate the mean distance between all points.
}
\details{
The mean distance of \eqn{p} points is calculated by the sum of the individual Euclidean distances
divided by the total numer of distances (given by \eqn{p*(p-1)/2}).
}

