% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca_mean_distance_grouping.R
\name{analysis_pca_mean_distance_grouping}
\alias{analysis_pca_mean_distance_grouping}
\title{Given a set of coordinates and a grouping variable, the mean distance is calculated for each group.}
\usage{
analysis_pca_mean_distance_grouping(MATRIX, FACTOR, factor_labs = NULL,
  nDim = 2, plot_mean_dist = T, Filename = "plot_mean_dist.pdf",
  plot_top_percent = NULL, ...)
}
\arguments{
\item{MATRIX}{- two column matrix with the point coordinates. Only the first two columns will be used.}

\item{FACTOR}{- used to split the data into groups. Character vector (or list of vetors) with the same length as rows
in \code{MATRIX}.}

\item{factor_labs}{- optional. Character vector of the length of \code{FACTOR} if \code{FACTOR} is a list or
length 1 if \code{FACTOR} is a character vector. Default (\code{NULL}; \code{FACTOR} names or letters used).}

\item{nDim}{- optional. Numeric vector of dimensions to use for the point coordinates. Default (\code{2}).}

\item{plot_mean_dist}{- logical. Should a plot of the mean distances be generated? Default (\code{TRUE}).}

\item{Filename}{- optional. Filename with path and extension. String added to distinguish \code{FACTOR} groups.
Default ("plot_mean_dist.pdf" saved to working directory).}

\item{plot_top_percent}{- optional. Numeric vector between 0 and 1 or integer. Default (NULL, i.e. not plotted).}

\item{...}{- further arguments for \code{plot_scatter()}.}
}
\value{
List containing a table of the mean distances (columns for group name, mean distance and size of group (N))
and a ggplot object of the data (NAs removed) if \code{plot_mean_dist} is TRUE.
}
\description{
Given a set of coordinates and a grouping variable, the mean distance is calculated for each group.
}
\details{
The mean distance of groups with one member (derived from the \code{FACTOR} labels) cannot be calculated.

\code{nDim} must be larger than 2.

\code{plot_top_percent} refers to the fraction or number of groups with the highest calculated mean distance
that should be plotted on a scatter plot. Items are recycled if not as many as the number of grouping factors is provided.
}
\seealso{
\link{analysis_pca_mean_distance_calculation}
}

