% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example_ECnumbers_vector}
\alias{data_example_ECnumbers_vector}
\title{Example of a vector containing Enzyme Commission (EC) numbers.}
\format{A character vector with 568 entries
\describe{
  \item{ECs}{1.1.1.10, 1.1.1.102, 1.1.1.105, 1.1.1.12, 1.1.1.14, 1.1.1.153, ...}
}}
\source{
\url{http://www.kegg.jp/kegg-bin/get_htext?eco00001} and expand the subsections to see the ECs.
}
\usage{
data_example_ECnumbers_vector
}
\description{
A dataset containing an example of a vector containing Enzyme Commission (EC) numbers
that can be used with the function \code{misc_geneVector_module()}
or by the function \code{query_genomes_to_modules()} after formatting it into a data frame (see function description).
}
\details{
When generating an EC number vector, make sure that all entries have the 4 nomenclature positions
(".-" denotes an unspecified field, e.g. "1.1.1.-").
}
\examples{
# Load data
data("data_example_ECnumbers_vector")
head(data_example_ECnumbers_vector)
# [1] "1.1.1.10"  "1.1.1.102" "1.1.1.105" "1.1.1.12"  "1.1.1.14"  "1.1.1.153"

}
\seealso{
\link{data_example_KOnumbers_vector}
}
\keyword{datasets}
