% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example_genomeIDs}
\alias{data_example_genomeIDs}
\title{Example of vector containing KEGG genome IDs to be used by \code{query_genomes_to_modules()}.}
\format{An object of class \code{character} of length 25.}
\usage{
data_example_genomeIDs
}
\description{
A dataset containing an example of a character vector
that can be used as the \code{GENOME_INFO} input to the function \code{query_genomes_to_modules()}.
}
\examples{
# Load data
data("data_example_genomeIDs")
head(data_example_genomeIDs)
# [1] "T04503" "T04203" "T03253" "T00526" "T00341" "T00552"

}
\seealso{
\link{plot_heatmap}
}
\keyword{datasets}
