% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example_KOnumbers_vector}
\alias{data_example_KOnumbers_vector}
\title{Example of a vector containing KEGG Ortholog (KO) identifiers}
\format{A character vector with 2896 entries
\describe{
  \item{KOs}{K00005, K00009, K00012, K00013, K00014, ...}
}}
\source{
\url{http://www.kegg.jp/kegg-bin/get_htext?eco00001} and expand the subsections to see the KOs.
}
\usage{
data_example_KOnumbers_vector
}
\description{
A dataset containing an example of a vector containing KEGG Orthologs (K numbers or KOs)
that can be used with the function function \code{misc_geneVector_module()}
or by the function \code{query_genomes_to_modules()} after formatting it into a data frame (see function description).
}
\examples{
# Load data
data("data_example_KOnumbers_vector")
head(data_example_KOnumbers_vector)
# [1] "K00005" "K00009" "K00012" "K00013" "K00014" "K00024"

}
\seealso{
\link{data_example_ECnumbers_vector}
}
\keyword{datasets}

