% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example_moduleIDs}
\alias{data_example_moduleIDs}
\title{Example of vector containing KEGG module IDs to be used by \code{query_genomes_to_modules()}.}
\format{An object of class \code{character} of length 6.}
\usage{
data_example_moduleIDs
}
\description{
A dataset containing an example of a character vector
that can be used as the \code{MODULE_ID} input to the function \code{query_genomes_to_modules()}.
}
\examples{
# Load data
data("data_example_moduleIDs")
head(data_example_moduleIDs)
# [1] "M00356" "M00357" "M00563" "M00567" "M00596" "M00377"

}
\seealso{
\link{plot_heatmap}
}
\keyword{datasets}
