% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example_multi_ECs_KOs}
\alias{data_example_multi_ECs_KOs}
\title{Example of a data frame with multiple datasets.}
\format{A data frame with the following columns:\preformatted{
 ID       - KEGG genome ID, T number (e.g. "T00001")
 ORG_ID   - KEGG organism ID,3-4 letter code (e.g. "eco")
 ORGANISM - Organism name (e.g. "Escherichia coli K-12 MG1655")
 KOs       - Concatenated string with the K numbers (e.g. "K00013;K00014;K00018;...").
 ECs       - Concatenated string with the EC numbers 
             (e.g. "1.1.1.1;1.1.1.100;1.1.1.130;...")
}}
\usage{
data_example_multi_ECs_KOs
}
\description{
A data frame containing an example of multiple datasets that could be analysed with the function \code{misc_geneVector_module()}
or by the function \code{query_genomes_to_modules()} after formatting it into a data frame (see function description).
The dataset entries are to illustrate what the KEGG genome data looks like.
}
\details{
The columns 'KOs' or 'ECs' need to be pointed at to do the module mapping and is specified by using the argument 'mapBy' in \code{query_genomes_to_modules()} and
The default is to use the column named 'KOs', but it can be named differently and specified in the argument field.

Specify the character to be used to split the string with the K/EC numbers using the argument 'split_vector_by' in \code{query_genomes_to_modules()}
}
\examples{
# Load data
data("data_example_multi_ECs_KOs")
head(data_example_multi_ECs_KOs)
}
\seealso{
\link{query_genomes_to_modules}
}
\keyword{datasets}

