% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example_sunburst}
\alias{data_example_sunburst}
\title{Example of the hierarchical data needed to generate a sunburst plot.}
\format{An object of class \code{data.frame} with 8 rows and 4 columns.}
\usage{
data_example_sunburst
}
\description{
A dataset containing an example of a data frame
that can be used with the function \code{plot_sunburst()}.
This function helps visualize hierarchical information by making concentric
"donut" plots which has the highest hierarchical level data at the centre going out.
}
\examples{
# Load data
data("data_example_sunburst")
head(data_example_sunburst)
#        CLASS_I    CLASS_II    CLASS_III    NAME_SHORT
# 1 Pathway module Carbohydrate \\& lipid metabolism Fatty acid metabolism    beta-Oxidation
# 2 Pathway module    Energy metabolism    Methane metabolism    Methanogenesis, from methanol
# 3 Pathway module    Energy metabolism    Methane metabolism    Methanogenesis, from acetate
# 4 Pathway module    Energy metabolism    Carbon fixation    Reductive acetyl-CoA pathway
# 5 Pathway module    Energy metabolism    Nitrogen metabolism    Dissimilatory nitrate reduction
# 6 Pathway module    Energy metabolism    Methane metabolism    Methanogenesis, from methylamine

}
\seealso{
\link{plot_sunburst}
}
\keyword{datasets}
