% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example_sunburst_fill_by}
\alias{data_example_sunburst_fill_by}
\title{Example of vector to provide alternative values to fill the outermost level of a sunburst plot.}
\format{An object of class \code{numeric} of length 8.}
\usage{
data_example_sunburst_fill_by
}
\description{
A dataset containing an example of a numerical vector
that can be used as the \code{fill_by} input to the function \code{plot_sunburst()}.
}
\examples{
# Load data
data("data_example_sunburst_fill_by")
head(data_example_sunburst_fill_by)
# M00087 M00356 M00357 M00377 M00530 M00563
#    0.0    1.0    0.8    0.0    0.0    0.5

}
\seealso{
\link{plot_sunburst}
}
\keyword{datasets}
