% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_module_shortName_mapping}
\alias{data_module_shortName_mapping}
\title{Data frame used to populate the short name in module_reference_table}
\format{A data frame with the following columns:\preformatted{
 ID         - KEGG module ID, M number (e.g. "M00001")
 NAME       - KEGG module name
 NAME_SHORT - Manually abbreviated KEGG module name (unique entries)
}}
\usage{
data_module_shortName_mapping
}
\description{
This data frame contains the following columns and is used to provide a manually abbreviated name to the
modules to ease plotting.
}
\examples{
# Load data
data("module_shortName_mapping")
head(module_shortName_mapping)
}
\seealso{
\link{parseKEGG_module}
}
\keyword{datasets}
