% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_check_duplicate_names.R
\name{misc_check_duplicate_names}
\alias{misc_check_duplicate_names}
\title{Check for duplicated strings and append letter code to distinguish them.}
\usage{
misc_check_duplicate_names(NAME_VECTOR, ...)
}
\arguments{
\item{NAME_VECTOR}{- character vector.}

\item{...}{- further arguments (currently unsupported)}
}
\description{
Check for duplicated strings and append letter code to distinguish them.
}
