% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_create_labels.R
\name{misc_create_labels}
\alias{misc_create_labels}
\title{Creates labels to add to a character vector to make all entries unique}
\usage{
misc_create_labels(N, ...)
}
\arguments{
\item{N}{- numeric. Indicates the length of the label vector to be return.}

\item{...}{- further arguments (currently unsupported)}
}
\value{
A character vector of length N with unique set of alpha-numeric labels.
}
\description{
Creates a label vector of length N, such that they are unique and help distinguish them
}
