% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_evaluate_block.R
\name{misc_evaluate_block}
\alias{misc_evaluate_block}
\title{Evaluate a KEGG module block definition.}
\usage{
misc_evaluate_block(gene_vector, BLOCK, KO_in_DEF_EC = FALSE)
}
\arguments{
\item{gene_vector}{- vector containing either K or EC numbers. See Details.}

\item{BLOCK}{- KEGG module DEFINITION BLOCK. See \link{parseKEGG_module}.}

\item{KO_in_DEF_EC}{- logical. If enzyme IDs are given, should lingering K numbers in the module DEFINITION
be assumed to be present? Default (\code{FALSE}).}
}
\description{
Evaluate a KEGG module block definition given a vector of genes (KEGG Orthologs -KOs-, identified with K number).
Some KOs can be mapped to Enzyme Commission (EC) numbers.
}
\details{
\code{gene_vector} must contain either K or EC numbers.
}
