% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_geneVector_module.R
\name{misc_geneVector_module}
\alias{misc_geneVector_module}
\title{Map a list of EC or K numbers to KEGG modules.}
\usage{
misc_geneVector_module(gene_vector, MODULE_ID, SEARCH_NAME, SEARCH_CLASS_I,
  SEARCH_CLASS_II, SEARCH_CLASS_III, EXCLUDE_NAME,
  use_module_reference_table = NULL, ...)
}
\arguments{
\item{gene_vector}{- character vector listing EC or K numbers. If more K numbers are given, the K number-based definition will be used. See Value.}

\item{MODULE_ID}{- optional. Character vector listing specific module IDs (e.g. M00001).}

\item{SEARCH_NAME}{- optional. Character vector listing terms to search in \code{NAME} field (case-insensitive).}

\item{SEARCH_CLASS_I}{- optional. Character vector listing terms to search in \code{CLASS_I} field (case-insensitive).}

\item{SEARCH_CLASS_II}{- optional. Character vector listing terms to search in \code{CLASS_II} field (case-insensitive).}

\item{SEARCH_CLASS_III}{- optional. Character vector listing terms to search in \code{CLASS_III} field (case-insensitive).}

\item{EXCLUDE_NAME}{- optional. Character vector listing terms that if matched in \code{NAME} field will be excluded (case-insensitive).}

\item{use_module_reference_table}{- optional. Provide a data frame with updated KEGG module database OR with custom-made modules.
Default (\code{NULL}; inbuilt data used). See Details.}

\item{...}{- further arguments such as \code{KO_in_DEF_EC}. See Details.}
}
\value{
List:
\code{$FRACTION} matrix containing the completeness fraction for the modules matched in the query along the columns.

\code{$METADATA} data frame with the metadata for the modules matched in the query.
 Columns:\preformatted{
	     (1)     MODULE_ID         - M number (e.g. M00001);
	     (2)     MODULE_NAME       - the KEGG given name for the module;
	     (3)     MODULE_NAME_SHORT - manually shortened module name (for plotting 
	                                  purposes);
	 (4 - 6)     CLASS_I - III     - hierarchical module classes;
	     (7)     DEFINITION        - KEGG module definition in terms of K or EC numbers 
	                                  (without optional K or EC numbers);
	     (8)     OPTIONAL          - the optional K or EC numbers that are part of the 
	                                  KEGG module definition (NA otherwise);
	     (9)     KOs_IN_DEF        - logical flag indicating whether there are K numbers 
	                                  involved in that module definition (if EC numbers 
	                                  are given);
}

\code{$ADD_INFO}  data frame with additional output information.
 Columns:\preformatted{
	 (1) MODULE_ID           - M number (e.g. M00001);
	 (2) MODULE_NAME_SHORT   - manually shortened module name (for plotting purposes);
	 (3) FRACTION            - the number of complete blocks divided by the total number
	                            of blocks;
	 (4) nBLOCKS             - the number of blocks that make up the KEGG module;
	 (6) COVERAGE            - the K numbers or ECs that are present and involved in the 
	                            KEGG module definition;
	 (7) OPTIONAL_PRESENT    - the K numbers or ECs that are present and are listed as
	                            optional;
}

See \link{parseKEGG_module} or visit \href{http://www.genome.jp/kegg/module.html}{http://www.genome.jp/kegg/module.html} for more information.
}
\description{
This function maps the list of Enzyme Commission (EC) numbers or KEGG orthologs (specified as K numbers) given by \code{gene_vector} to the
KEGG modules using an in-build reference table (\code{module_reference_table}) of all the KEGG modules in the database at the time of release.
}
\details{
The modules to be analysed are determined by search queries.
These can search across the module NAME, CLASS (I-III) or by specifying the module ID (M number). See the Argument section.
When none of the optional arguments are specified, the default is to search all modules.
An additional argument, \code{EXCLUDE_NAME}, can be used to exclude modules with a certain term in the \code{NAME} field.
For example, specifying \code{EXCLUDE_NAME} = "biosynthesis" would return the search across all modules,
except those that contain "biosynthesis" in the name.
For instance, module "M00005" named "PRPP biosynthesis, ribose 5P => PRPP" and 168 others would be excluded form the search.

If neither \code{MODULE_ID} or at least one \code{SEARCH} term is specified, the default is to analyse all modules.

The \code{use_} argument allow users with KEGG FTP access to provide the updated data from the KEGG databases in the form of 
reference tables AND/OR for advanced users to provide custom-made modules (see below).
These reference tables can be generated with \code{parseKEGG_module} and need to have a specific format (see function descriptions for details on format).

The module definition (contained in \code{module_reference_table}) describes the relationship between genes and modules
and is used to identify the modules in which \code{gene} is involved.
The user can provide custom-made module definitions that use the logical expression format (however, the table format must be conserved!).
}
\examples{
output_module_ECs <- misc_geneVector_module(data_exampleECnumbers_vector,
                                             MODULE_ID=paste("M0000",1:5,sep=""))

output_module_KOs <- misc_geneVector_module(data_exampleKOnumbers_vector,
                                             MODULE_ID=paste("M0000",1:5,sep=""))
}
\seealso{
\link{parseKEGG_module}
}

