% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_module_definition_block_EC.R
\name{misc_module_definition_block_EC}
\alias{misc_module_definition_block_EC}
\title{KEGG module definition processing: subblocks}
\usage{
misc_module_definition_block_EC(BLOCK, ORTHOLOGS, ...)
}
\arguments{
\item{BLOCK}{- string containing a block of a KEGG module DEFINITION (logical expression using K numbers).}

\item{ORTHOLOGS}{- vector listing the K numbers and related EC numbers.}

\item{...}{- further arguments (currently unsupported)}
}
\description{
Format the KEGG module database - Format the KEGG module DEFINITION for ease of analysis by excluding optional KEGG orthologs.
Used by \code{parseKEGG_module()}.
}
\details{
The KEGG module definition uses optional KEGG orthologs, indicated by a "-". These are removed and the corresponding EC number stored.
}
\seealso{
\link{parseKEGG_module}
}
