% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_module_definition_check.R
\name{misc_module_definition_check}
\alias{misc_module_definition_check}
\title{KEGG module definition processing: block formatting.}
\usage{
misc_module_definition_check(DEFINITION, ...)
}
\arguments{
\item{DEFINITION}{- string containing a KEGG module DEFINITION (logical expression using K numbers).}

\item{...}{- further arguments (currently unsupported)}
}
\description{
Format the KEGG module DEFINITION for ease of analysis.
}
\details{
The KEGG module definition uses both spaces and plus signs to indicate \code{'AND'} operations.
However, the \code{'AND'} operation can be used to split the module definition into BLOCKS or to indicate
molecular complex composition.
To simplify analysis, we will use spaces ONLY to delimit KEGG module BLOCKS and the plus sign ONLY to
indicate molecular complexes and \code{'AND'} operations within blocks.
}
\seealso{
\link{parseKEGG_module}
}
