% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_module_definition_optional.R
\name{misc_module_definition_optional}
\alias{misc_module_definition_optional}
\title{KEGG module definition processing: optional KEGG ortholog exclusion}
\usage{
misc_module_definition_optional(BLOCK, ORTHOLOGS)
}
\arguments{
\item{BLOCK}{- string containing a block of a KEGG module DEFINITION (logical expression using K numbers).}

\item{ORTHOLOGS}{- vector listing the K numbers and related EC numbers.}
}
\value{
A list with the formatted block (\code{$thisBlock}) and any optional K and EC numbers (\code{$thisOptional_KO} and \code{$thisOptional_EC})
}
\description{
Format the KEGG module database - Format the KEGG module definition for ease of analysis by excluding optional KEGG orthologs.
Used by \code{parseKEGG_module()}.
}
\details{
The KEGG module definition uses optional KEGG orthologs, indicated by a "-". These are removed and the corresponding EC number stored.
}
\seealso{
\link{parseKEGG_module}
}
