% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_module_subgroup_indexing.R
\name{misc_module_subgroup_indexing}
\alias{misc_module_subgroup_indexing}
\title{Subgroup check - formatting KEGG module definition.}
\usage{
misc_module_subgroup_indexing(DEFINITION)
}
\arguments{
\item{DEFINITION}{- KEGG module definition}
}
\description{
This functions helps format the KEGG module DEFINITION by checking all the bracket-delimitted BLOCKS
to then remove flanking brackets (unnecessary).
}
\seealso{
\link{parseKEGG_module},\link{misc_module_definition_check}
}

