% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_compound.R
\name{parseKEGG_compound}
\alias{parseKEGG_compound}
\title{Parse the KEGG compound database}
\usage{
parseKEGG_compound(KEGG_path, outDir = "output", verbose = T, ...)
}
\arguments{
\item{KEGG_path}{- string pointing to the location of the KEGG database parent folder.}

\item{outDir}{- string pointing to the output folder. Default ("output/"). \code{NULL} overwrites writting files.}

\item{verbose}{- logical. Should progress be printed to the screen? Default (\code{TRUE}).}

\item{...}{- other arguments for \code{parseKEGG_file()}.}
}
\value{
Generates compound_reference_table (.txt & .rda; saved to \code{'outDir'}) and returns
a data frame with as many rows as entries and the following columns (or variables):
\preformatted{
	 (1) ID         - C number identifier (e.g. "C00001");
	 (2) NAME       - compund name(s);
	 (3) FORMULA    - chemical formula;
	 (4) EXACT_MASS - compound's mass;
	 (5) MOL_WEIGHT - molecular weight;
	 (6) REMARK     - relationship with D number and others;
	 (7) REACTION   - reactions IDs (R#####) in which the compound is involved;
	 (8) PATHWAY    - pathway(s) in which the compound is involved (map### and name);
	 (9) MODULE     - module(s)  in which the compound is involved (M##### and name);
	 (10) ENZYME    - enzyme IDs (#.#.#.#) catalysing a reaction in which the compound is involved;
	 (11) BRITE;    (12) DBLINKS;   (13) ATOM;      (14) BOND;  (15) COMMENT;
	 (16) BRACKET;  (17) SEQUENCE;  (18) REFERENCE;
}
In all instances, multiple entries in a given column are separated by '[;]'.
}
\description{
Read the KEGG compound database text file ("\code{KEGG_path}/ligand/compound.tar.gz" OR "\code{KEGG_path}/ligand/compound/compound")
and the organism information and format it into a reference table.
Note that the columns are automatically generated by the \code{parseKEGG_file} function and
columns or variables specific to the KEGG compound database are further formated.
}
\seealso{
\link{parseKEGG_file}
}

