% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_enzyme.R
\name{parseKEGG_enzyme}
\alias{parseKEGG_enzyme}
\title{Parse the KEGG enzyme database}
\usage{
parseKEGG_enzyme(KEGG_path, outDir = "output", verbose = T, ...)
}
\arguments{
\item{KEGG_path}{- string pointing to the location of the KEGG database parent folder. The path to the required file is contained within the function.}

\item{outDir}{- string pointing to the output folder. Default ("output/").}

\item{verbose}{- logical. Should progress be printed to the screen? Default (\code{TRUE})}

\item{...}{- other arguments for \code{parseKEGG_file()}.}
}
\value{
Generates enzyme_reference_table (.txt & .rda; saved to \code{'outDir'}) and returns
a data frame with as many rows as entries and the following columns (or variables):
\preformatted{
	 (1) ID         - Enzyme Commission (EC) number (e.g. "1.1.1.1"; 4 positions);
	 (2) NAME       - enzyme name(s);
	 (3) CLASS_I    - enzyme class; refers to the first position.
	                   CLASSES:
	                    1. Oxidoreductases, 2. Transferases, 3. Hydrolases,
	                    4. Lyases,          5. Isomerases,   6. Ligases
	 (4) CLASS_II   - further enzyme class info; refers to the second position
	                   (different for every class);
	 (5) CLASS_III  - further enzyme class info; refers to the third  position
	                   (different for every class);
	 (6) SYSNAME    - alternative names;
	 (7) REACTION   - reaction(s) the enzyme catalyses;
	 (8) ALL_REAC   - reaction ID(s) (R number);
	 (9) SUBSTRATE  - substrate name and ID(s);
	 (10) PRODUCT   - product   name and ID(s);
	 (11) COMMENT; 	   (12) HISTORY; 	 (13) REFERENCE;
	 (14) PATHWAY   - pathway(s) in which the enzyme is involved (ec### and name);
	 (15) ORTHOLOGY - related KEGG Ortholog(s) (K number; K##### and name);
	 (16) GENES; 	     (17) DBLINKS;

 *In all instances, multiple entries in a given column are separated by '[;]'.
}
}
\description{
Read the KEGG enzyme database text file and format it into a reference table.
}
\details{
The columns are automatically generated by the \code{parseKEGG_file} function into variables,
which are further formatted specifically for the KEGG enzyme database.

The text file used is "\code{KEGG_path}/ligand/enzyme/enzyme".

It decompresses "\code{KEGG_path}/ligand/enzyme.tar.gz" if needed.
}
\examples{

KEGG_path <- "~/KEGG" # MODIFY TO KEGG PARENT FOLDER!
# The parent folder should contain the following (KEGG FTP structure):
#	brite/
#	genes/
#	ligand/
#	medicus/
#	module/
#	pathway/
#	README.kegg
#	RELEASE
#	xml/

enzyme_reference_table <- parseKEGG_enzyme(KEGG_path)
 # A .txt file (tab separated) is written to output/ (relative to current working directory)

}
\seealso{
\link{parseKEGG_file}
}
