% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_file.R
\name{parseKEGG_file}
\alias{parseKEGG_file}
\title{Parse any KEGG file without extension}
\usage{
parseKEGG_file(FILE_PATH, split_pattern = "ENTRY", pathway_trim = T,
  verbose = T)
}
\arguments{
\item{split_pattern}{- string to use to identify start of new section/entry. Default ("ENTRY").}

\item{pathway_trim}{- logical. Should the file 'KEGG_path/pathway/pathway' be trimmed to only include 'map' entries? (i.e. exclude ko, ec and organism-specific pathways).
Default (\code{TRUE}).}

\item{verbose}{- logical. Should progress be printed to the screen? Default (\code{TRUE}).}

\item{FILE}{- string pointing to the location of the file WITHOUT EXTENSION (uncompressed).}
}
\value{
A data frame with the formatted data.See the file-specific functions.
}
\description{
Read the KEGG database text file without extension (e.g. 'module', 'enzyme', 'genome') and format it into a
reference table.
Generates DATABASE_reference_table (data frame).
}
\details{
File must be decompressed before being processed.
The functions that use this function (listed below in \code{see also}) perform this step if necessary.
}
\seealso{
\link{parseKEGG_compound},\link{parseKEGG_enzyme},\link{parseKEGG_genome}, \link{parseKEGG_module},

\link{parseKEGG_reaction}, \link{parseKEGG_execute_all}
}

