% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_genome.R
\name{parseKEGG_genome}
\alias{parseKEGG_genome}
\title{Parse the KEEG genome database}
\usage{
parseKEGG_genome(KEGG_path, outDir = "output", addECs = T, addKOs = T,
  includeVirues = F, verbose = T, ...)
}
\arguments{
\item{KEGG_path}{- string pointing to the location of the KEGG database parent folder.}

\item{outDir}{- optional. String pointing to the output folder. Default ("output/"). \code{NULL} overwrites writting files.}

\item{addECs}{- logical. Should the list of ECs be retrieved for each organism? Default (\code{TRUE}).}

\item{addKOs}{- logical. Should the list of K numbers be retrieved for each organism? Default (\code{TRUE}).}

\item{includeVirues}{- logical. Should viral genomes be included? Default (\code{FALSE}).}

\item{verbose}{- logical. Should progress be printed to the screen? Default (\code{TRUE}).

 @param ...       - other arguments for \code{parseKEGG_file()}.}
}
\value{
Generates genome_reference_table (.txt & .rda; saved to \code{'outDir'}) and returns
a data frame with as many rows as entries and the following columns (or variables):
\preformatted{
	 (1) ID                   - KEGG genome   identifier (T0 number; e.g. "T00001");
	 (2) ORG_ID               - KEGG organism identifier (3 or 4 letter code; e.g. "hin");
	 (3) STATUS               - genome sequence status (e.g. "Complete Genome");
	 (4) NAME                 - various identifiers (e.g. "hin, HAEIN, 71421");
	 (5) ORGANISM             - organism name (Genus species sp);
	 (6) ANNOTATION           - type of annotation (one of "manual", "KOALA" or  "none");
	 (7) TAXONOMY;             (8)  DATA_SOURCE;
	 (9) ORIGINAL_DB          - original database;
	 (10) KEYWORDS;            (11) DISEASE;
	 (12) COMMENT;             (13) CHROMOSOME;
	 (14) STATS_N_NUCLEOTIDES - statistics, number of nucleotides;
	 (15) STATS_N_GENES_PROT  - statistics, number of protein-encoding genes;
	 (16) STATS_N_GENES_RNA   - statistics, number of RNA-encoding genes;
	 (17) REFERENCE           - reference(s) for study from which genomic sequence was derived;
	 (18) PLASMID;             (19) DBLINKS;
	 (20) KOs                 - concatenated string of KEGG Orthologs (K numbers; e.g. "K00001");
	 (21) ECs                 - concatenated string of Enzyme Classification (EC) numbers (e.g. "1.1.1.1").

	    *In all instances, multiple entries in a given column are separated by '[;]'.
}
}
\description{
Read and format the KEGG genome database text file ("\code{KEGG_path}/genes/genome.tar.gz" OR
"\code{KEGG_path}/genes/genome/genome")
and the organism information and format it into a reference table.
Note that the columns are automatically generated by the \code{parseKEGG_file} function and
columns or variables specific to the KEGG genome database are further formated.
}
\seealso{
\link{parseKEGG_file}
}

