% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_ko.R
\name{parseKEGG_ko}
\alias{parseKEGG_ko}
\title{Parse the KEGG orthology (KO) database}
\usage{
parseKEGG_ko(KEGG_path, outDir = "output", verbose = T, ...)
}
\arguments{
\item{KEGG_path}{- string pointing to the location of the KEGG database parent folder.}

\item{outDir}{- optional. String pointing to the output folder. Default ("output/"). \code{NULL} overwrites writting files.}

\item{verbose}{- logical. Should progress be printed to the screen? Default (\code{TRUE}).}

\item{...}{- other arguments for \code{parseKEGG_file()}.}
}
\value{
Generates ko_reference_table (.txt & .rda; saved to \code{'outDir'}) and returns
a data frame with as many rows as entries and the following columns (or variables):
\preformatted{

	    *In all instances, multiple entries in a given column are separated by '[;]'.
}
}
\description{
Read and format the KEGG orthology (KO) database (containing ortholog (gene) information) text file ("\code{KEGG_path}/genes/ko.tar.gz" OR
"\code{KEGG_path}/genes/ko/ko") into a reference table.
}
\details{
Note that the columns are automatically generated by the \code{parseKEGG_file} function and
columns or variables specific to the KEGG ortholog database are further formated.
}
\seealso{
\link{parseKEGG_file}
}

