% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_ko_reaction.R
\name{parseKEGG_ko_reaction}
\alias{parseKEGG_ko_reaction}
\title{Map KEGG orthologs (KOs) to Reaction IDs}
\usage{
parseKEGG_ko_reaction(KEGG_path, outDir = "output", verbose = T, ...)
}
\arguments{
\item{KEGG_path}{- string pointing to the location of the KEGG database parent folder. The path to the required file is contained within the function.}

\item{outDir}{- string pointing to the output folder. Default ("output/").}

\item{verbose}{- logical. Should progress be printed to the screen? Default (\code{TRUE}).}

\item{...}{- further arguments for \code{parseKEGG_file.list()}.}
}
\value{
Data frame establishing the relationship between KO numbers and reactions (R number) (binary).
\preformatted{
> ko_reaction_map[1:3,1:3]

              R00005 R00006 R00008
    K00001      0      0      0
    K00002      0      0      0
    K00003      0      0      0
}
}
\description{
Map KEGG orthologs (KOs) to Reaction IDs and format it into a matrix with binary indicator for mapping/relationship.
Generates 'ko_reaction_map' (.txt & .rda).
}
\examples{
KEGG_path     <- "~/KEGG" # MODIFY TO KEGG PARENT FOLDER!
# The parent folder should contain the following (KEGG FTP structure):
#	brite/
#	genes/
#	ligand/
#	medicus/
#	module/
#	pathway/
#	README.kegg
#	RELEASE
#	xml/

ko_reaction_map <- parseKEGG_ko_reaction(KEGG_path)
 # A .txt file (tab separated) is written to output/ (relative to current working directory)

}
\seealso{
\link{parseKEGG_file.list}
}
