% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_module.R
\name{parseKEGG_module}
\alias{parseKEGG_module}
\title{Parse the KEGG module database}
\usage{
parseKEGG_module(KEGG_path, outDir = "output", verbose = T,
  shortName_file_path = "", ...)
}
\arguments{
\item{KEGG_path}{- string pointing to the location of the KEGG database parent folder}

\item{outDir}{- string pointing to the output folder. Default ("output/"). \code{NULL} overwrites existing files.}

\item{verbose}{- logical. Should progress be printed to the screen? Default (\code{TRUE})}

\item{shortName_file_path}{- file path to table containing Module IDs (column 1), Short name (column 2). Default (\code{""})}

\item{...}{- other arguments for \code{parseKEGG_file()}}
}
\value{
Generates module_reference_table (.txt & .rda; saved to \code{'outDir'}) and returns
a data frame with as many rows as entries and the following columns (or variables):
\preformatted{
	      (1) ID             - KEGG module identifier (M number; e.g. "M00001");
	      (2) NAME           - KEGG module name;
	      (3) NAME_SHORT     - abbreviated module name (for visualization purposes);
	      (4) DEFINITION_KOs - module definition as a logical expression (see Details) 
	                            in terms of KEGG Orthologs;
	      (5) DEFINITION_ECs - module definition as a logical expression (see Details) 
	                            in terms of EC numbers;
	      (6) OPTIONAL_KOs   - optional K numbers listed;
	      (7) OPTIONAL_ECs   - optional EC numbers listed;
	      (8) ORTHOLOGY      - relationship between K and EC numbers;
	 (9 - 11) CLASS_I - III  - hierarchical module classes;
	     (12) PATHWAY        - pathway(s)  in which the module is involved (map### and 
	                            name);
	     (13) REACTION       - reaction(s) in which the module is involved (R##### and 
	                            their corresponding compound IDs);
	     (14) COMPOUND       - compund(s)  in which the module is involved (C#####);
	     (15) DBLINKS;         (16) RMODULE;         (17) REFERENCE;         
	     (18) COMMENT;         (19) BRITE;

 *In all instances, multiple entries in a given column are separated by '[;]'.
}
}
\description{
Read the KEGG module database text file and format it into a reference table.
}
\details{
The columns are automatically generated by the \code{parseKEGG_file} function into variables, 
which are further formatted specifically for the KEGG module database.

The text file used is "\code{KEGG_path}/module/module". 

It decompresses "\code{KEGG_path}/module/module.gz" if needed.

DEFINITION: Logical Expression (adapted from \link{http://www.genome.jp/kegg/module.html})

The MODULEs (identified by an M number; e.g. "M00001") are defined by a logical expression (DEFINITION) of KEGG orthologs
(KO numbers, KOs; e.g. "K00001") and sometimes other M numbers, facilitating the automatic evaluation of whether a
module is complete in a given genome.

A MODULE is made up of BLOCKS (SPACE delimited). Each block is defined by a logical expression
to determine which KOs are needed in the definition.
All BLOCKS must be present to be able to state that a MODULE is COMPLETE.

The KEGG module DEFINITION has been formatted to simplify its use, but the logical expression is conserved.
Where space or a plus sign represent AND operations in the KEGG definition,
we have replaced all instances with \code{'&'}. Similarly, we have replaced all comas (used within KEGG to represent an OR operation)
with pipes (\code{'|'}).

We have also translated the K number based DEFINITION to an enzyme based DEFINITION using the ORTHOLOGY information.
K numbers can be mapped to enzymes using the Enzyme Classification (EC) numbers (redundancy expected).
Note that not all K numbers have an association to an EC number.
In these instances, the K numbers have been left in the MODULE DEFINITION.
When evaluating whether a MODULE is complete using EC numbers, the user must decide whether to assume that those genes
are present or not (deafult). See \link{query_genomes_to_modules}

Optional items in the complex or definition (denoted by a minus sign) have been removed from the main definition
and listed as a separate column under \code{'OPTIONAL_' (EC or K numbers)}.

\code{shortName_file_path} can be used to provide the path to a file listing:\preformatted{
 Column named 'ID'         - KEGG module ID (M numbers),
 Column named 'NAME'       - KEGG module NAME,
 Column names 'NAME_SHORT' - manually abbreviated names for plotting purposes.
}
\code{NAME_SHORT} abbreviations: 2-CRS, two-component regulatory system; PS, photosystem/photosynthesis; pwy, pathway; TS, transport system; R, resistance;
}
\seealso{
\link{parseKEGG_file}, \link{misc_module_definition_check},

\link{misc_module_definition_optional},' \link{misc_module_definition_block_EC}
}

