% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_process_KEGG_taxonomy.R
\name{parseKEGG_process_KEGG_taxonomy}
\alias{parseKEGG_process_KEGG_taxonomy}
\title{Retrieve the taxonomy listed as part of the KEGG genome database.}
\usage{
parseKEGG_process_KEGG_taxonomy(genome_reference_table,
  taxonomy_header = "TAXONOMY", org_header = 1)
}
\arguments{
\item{genome_reference_table}{- table containing the different genome entires (rows) and data (cols). See \link{parseKEGG_genome}.}

\item{taxonomy_header}{- optional. Character with header name or number indicating column index for taxonomic information. Default ("TAXONOMY").}

\item{org_header}{- optional. Character with header name or number indicating column index for organism name or ID. Use to name the output rownames. Default (1; first column).}
}
\value{
Data frame with 6 columns containing the taxonomic information (columns: KINGDOM, PHYLUM, CLASS, ORDER, FAMILY, GENUS).
The rownames contain the organism information (name or ID) as specified with \code{org_header}.
}
\description{
Retrieve the taxonomy listed as part of the KEGG genome database.
}
\details{
After processing the KEGG genome database, this function can extract the taxonomic information.
This information should specify KINGDOM, PHYLUM, CLASS, ORDER, FAMILY, GENUS (6).
NOTE: it is incomplete and inconsistent, most likely because it is derived from multiple sources.
}
\seealso{
\link{parseKEGG_genome}
}

