% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot}
\usage{
plot_scatter(plot_data, X = 1, Y = 3, colBy = NULL, xLab = "",
  yLab = "", Filename = "plot_scatter.pdf", xLabs_angle = T, Width = 24,
  Height = 18)
}
\arguments{
\item{plot_data}{- data frame. The columns to be plotted are indicated with \code{X} and \code{Y}.Order given is conserved.}

\item{X, Y}{- optional. Number indicating column to use for x axis and y axis, respectively. Default (1 and 3).}

\item{colBy}{- optional. Number indicating column to use for coloring grouping. Default (\code{NULL}).}

\item{xLab, yLab}{- optional. String to use for x label and y label, respectively. Default 
(first and second column names, respectively).}

\item{Filename}{- optional. String(s) to use for file name. Default ("plot_scatter.pdf"). If set to \code{""} a file is not written.}

\item{Width}{- width for figure file. Default (12in).}

\item{Height}{- height for figure file. Default (8in).}
}
\value{
ggplot2 plot object
}
\description{
Scatter plot for pairs of categorical data with a numeric value.
}
\details{
This function is used by \link{analysis_genomes_module_output()} to plot the module variance accross genomes and

by \link{analysis_pca_mean_distance_grouping()} to plot the mean PCA distance of the groups analysed.

If \code{colBy} is set to \code{NULL}, no grouping will be done for colouring (a scale warning will be issued that can be safely ignored).
}
\examples{
plot_data_example <- data.frame("Groups"=LETTERS[1:12],
                                "Factor"=c(rep(1,4),rep(2,4),rep(3,4)),
                                "Value"=runif(12,-10,10),stringsAsFactors = FALSE)

plot_scatter(plot_data_example,Filename = "")

# Change plot order to be according to the numeric value
plot_data_example <- plot_data_example[order(plot_data_example$Value),]
plot_scatter(plot_data_example,Filename = "")

}
\seealso{
\link{analysis_genomes_module_output},\link{analysis_pca_mean_distance_grouping}
}

