% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_byFactors.R
\name{plot_scatter_byFactors}
\alias{plot_scatter_byFactors}
\title{Scatter plot with overlapping factors/groups.}
\usage{
plot_scatter_byFactors(MATRIX, FACTOR, factor_labs = NULL,
  Filename = "plot_scatter_byFactors.pdf", Width = 9, Height = 7)
}
\arguments{
\item{MATRIX}{- two column matrix to plot. Only the first two columns will be used.}

\item{FACTOR}{- character vector or list of character vectors used to split the data into groups.}

\item{factor_labs}{- optional. Character vector to distinguish \code{FACTOR} groups. Default (\code{NULL}). See Details.}

\item{Filename}{- optional. Character vector containing the file path(s) to save image(s) WITH extension.
The device to save is determined by file extension. Default  ("plot_scatter_byFactors.pdf").
If set to \code{""}, no file will be written.}
}
\value{
A list with as many entires as factors is generated (one for each factor) using \code{factor_labs} for the names.
For every \code{FACTOR}, a list will contain:
\preformatted{
 $nGroups - numeric. The number of groups found for that FACTOR (NOTE: That entries with "" will be excluded.)
 $table   - data frame with the number of entries for each group found.
 $file    - character vector of file name(s).
}
}
\description{
Represent different groups as defined by FACTOR(S) in a scatter plot.
This function is used by \link{analysis_genomes_module_output()} to plot the first two Principal Components (PCs) from the PCA analysis,
overlapping different factors or groups (as determined by \code{FACTOR}).
It is also used by \link{analysis_pca_mean_distance_grouping()} to highlight a single group on PC plot.
}
\details{
\code{factor_labs} is used as an extension for the filename for the plot files.
The names of the \code{FACTOR} object OR "factor" followed by a number
 will be used if \code{factor_labs} is not specified (i.e. \code{factor_labs = NULL}).
}
\seealso{
\link{analysis_genomes_module_output},\link{analysis_pca_mean_distance_grouping}
}

