% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sunburst.R
\name{plot_sunburst}
\alias{plot_sunburst}
\title{Sunburst plot}
\usage{
plot_sunburst(DATA, centerLabel = "", ANGLE = FALSE, fill_by = NULL,
  fill_by_mean = FALSE, outer_text = TRUE,
  outer_text_levelN_minus_1 = TRUE, legend_name = "Counts\\n",
  sunburst = TRUE, setMax = NULL, setMin = NULL, textSize = NULL,
  textColour = "black", Filename = "", WIDTH = 25, HEIGHT = 25, ...)
}
\arguments{
\item{DATA}{- Data frame containing hierarchical data by columns, where the left column is the highest level and the right column the lowest one.}

\item{centerLabel}{- optional. String to be used for centre label. Default (\code{""}; i.e. no text).}

\item{ANGLE}{- optional. Should the angle of the text be adjusted to the position where it's at? Default (\code{FALSE}; i.e. horizontal text).}

\item{fill_by}{- optional. Numeric vector containing values to be used to fill outer most ring (see Details). Default (\code{NULL}).}

\item{fill_by_mean}{- optional. Should the categories in the inner rings be filled (colored) by the mean of the fill_by value provided. Default (\code{FALSE}).}

\item{outer_text}{- optional. Should the text from the lowest level (right column) be included? Default (\code{TRUE}).}

\item{outer_text_levelN_minus_1}{- optional. Should the text from the second lowest level (second column from right) be included? Default (\code{TRUE}).}

\item{legend_name}{- optional. Default ("Counts\\n"; i.e. have space between name and legend bar).}

\item{sunburst}{- optional. Should the sunburst be made (default) or should the output be bars? Default (\code{TRUE,}).}

\item{setMax}{- optional. Numerical value to set the maximum limit. Categories with higher values than those specified will be excluded, making them light blue.  Default (\code{NULL}).}

\item{setMin}{- optional. Numerical value to set the minimun limit. Categories with lower  values than those specified will be excluded, making them light blue.  Default (\code{NULL}).}

\item{textSize}{- optional. Numerical value to set the text size. Default (\code{NULL}).}

\item{textColour}{- optional. String to set the text colour. Default ("black").}

\item{Filename}{- optional. A character vector containing the file path to save image(s). The device to save is determined by file extension.
Default (\code{""}; i.e. printed to device, file not written).}

\item{WIDTH}{- optional. Width  size for file. Note that text size in plot scales with \code{WIDTH}. Default (25 in).}

\item{HEIGHT}{- optional. Height size for file. Default (25 in).}
}
\value{
List containing the \code{$DATA} used to generate the plot and the plot itself as a ggplot object \code{$p}.
}
\description{
This function arranges the hierarchical data and fills based on the counts for each category and level.
}
\details{
This function arranges a hierarchical data set into a dart-style chart, which is then partitioned and colored based on the data structure.
The outer most section displays the number of elements (counts) in the lowest (most specific) hierarchical level.
The user has the option to fill the outer level with user specified values given to \code{fill_by}
or to fill with the mean value \code{fill_by_mean}. See \code{Details}.


The sunburst is generated by having a ring containing the hierarchical data from highest level (most generic) at the centre,
followed by concentric rings that go out as the data becomes more specific.
The outer-most ring is colored (filled) by the number of members (counts) in that category.
The hierarchy is conserved throughout the data levels.

\preformatted{
LEVEL_1   LEVEL_2   LEVEL_3
      a       aa       aa1
      a       aa       aa2
      a       ab  		  ab1
      a       ab  		  ab2
      b       ba  		  ba1
      b       ba  		  ba2
      b       bb  		  bb1
      b       bb  		  bb2
      c       ca  		  ca1
      c       ca  		  ca2
      c       cb  		  cb1
      c       cb  		  cb2
}

The above data would result in a sunburst plot that would have 3 categories in the inner-most ring, each with 2 categories in the second ring (6 total)
and 12 categories in the outer ring, one for each of the second level categories.
Therefore, the counts would all be 1 (see the first example for a case with different levels).

\code{fill_by} allows the user to display other values associated with the outer-most categories, such as the module fraction completeness (mfc), rather than the membership count.
If \code{fill_by_mean} is set to \code{TRUE}, then the inner rings are also filled (colored) based on the mean of the \code{fill_by} value of the data making up the category.

\code{fill_by} should contain as many entries as there are categories at the last level.
If more are provided, only the required entries will be used
(i.e. if 5 values are provided for data with 3 categories, only the first 3 values will be used).
Only as many entries as there are rows in \code{DATA} are allowed.

Function developed by expanding and modifying Yahia El Gamal's blog post "Create Basic Sunburst Graphs with ggplot2"
([https://medium.com/optima-blog/create-basic-sunburst-graphs-with-ggplot2-7d7484d92c61])
}
\examples{
data(data_example_sunburst)

#  CLASS_I            CLASS_II            CLASS_III            NAME_SHORT
#	Pathway module    Carbohydrate & lipid metabolism    Fatty acid metabolism    beta-Oxidation
#	Pathway module        Energy metabolism    Methane metabolism    Methanogenesis, from methanol
#	Pathway module        Energy metabolism    Methane metabolism    Methanogenesis, from acetate
#	Pathway module        Energy metabolism    Carbon fixation        Reductive acetyl-CoA pathway
#	Pathway module        Energy metabolism    Nitrogen metabolism    Dissimilatory nitrate reduction
#	Pathway module        Energy metabolism    Methane metabolism    Methanogenesis, from methylamine
#	Pathway module        Energy metabolism    Methane metabolism    Methanogenesis, from CO2
#	Pathway module        Energy metabolism    Sulfur metabolism    Dissimilatory sulfate reduction

# Simplest plot using count data.
plot_data <- plot_sunburst(sunbust_data,WIDTH = 8) # WIDTH scales text size

# Specify values to be used for outer ring (lowest level) and change legend name accordingly

data(data_example_sunburst_fill_by)
        # mcf for the modules associated with the data shown above for an example dataset.
plot_data <- plot_sunburst(data_example_sunburst,centerLabel = "Org A",
                           fill_by = data_example_sunburst_fill_by,outer_text = F,WIDTH = 8,
                           legend_name = "fill_by")

# Also fill inner rings (levels) according to the mean values determined by fill_by.
plot_data <- plot_sunburst(data_example_sunburst,centerLabel = "Org A",
                           fill_by = data_example_sunburst_fill_by,fill_by_mean = T,
                           outer_text = F,WIDTH = 8,legend_name = "fill_by")

}
