% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_genes_to_genomes.R
\name{query_genes_to_genome}
\alias{query_genes_to_genome}
\title{Find the genome(s) that express a set of genes.}
\usage{
query_genes_to_genome(genes, use_genome_reference_table = NULL,
  use_ko_reference_table = NULL, use_enzyme_reference_table = NULL)
}
\arguments{
\item{genes}{- character vector of KO or EC numbers.}

\item{use_genome_reference_table}{- optional. Provide a data frame with updated KEGG genome database.
Default (\code{NULL}; inbulit data used). See Details.}

\item{use_ko_reference_table}{- optional. Provide a data frame with updated KEGG ortholog database.
Default (\code{NULL}; inbulit data used). See Details.}

\item{use_enzyme_reference_table}{- optional. Provide a data frame with updated KEGG enzyme database.
Default (\code{NULL}; inbulit data used). See Details.}
}
\value{
Data frame containing the \code{genes} (rows, specified in the rownames) and the T number of the KEGG genomes (columns) with a binary indicator for presence of gene in given genome.
}
\description{
The genes can be either an enzyme, given by its Enzyme Classification (EC) number (e.g. "1.1.1.1"),
or a KEGG ortholog identifier (K number, e.g. "K00001").
}
\details{
The \code{use_} set of arguments allow users with KEGG FTP access to provide the updated data from the KEGG databases in the form of reference tables AND/OR for advanced users to provide contum-made modules (see below).
These reference tables can be generated with the \code{parseKEGG} family of functions and need to have a specific format (see function descriptions for details on format).

If providing \code{use_genome_reference_table}, make sure that the \code{parseKEGG_genome} function is run with arguments \code{addECs = T} and/or \code{addKOs = T}
to include genes that make up the genomes.

\code{ko_reference_table} and \code{enzyme_reference_table} are used to verify that the \code{genes} provided exist within the KEGG database.
}
\examples{
genomes <- query_genes_to_genome("K00844")

genomes <- query_genes_to_genome("1.1.1.1")

genomes <- query_genes_to_genome(genes = paste("K0000",1:3,sep=""))

}
\seealso{
\link{parseKEGG_genome}, \link{ko_reference_table}, \link{enzyme_reference_table}
}

