% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_genes_to_modules.R
\name{query_genes_to_modules}
\alias{query_genes_to_modules}
\title{Given a set of genes, find the modules they are involved in.}
\usage{
query_genes_to_modules(genes, use_module_reference_table = NULL,
  use_ko_reference_table = NULL, use_enzyme_reference_table = NULL)
}
\arguments{
\item{genes}{- character vector (length 1). K or EC number.}

\item{use_module_reference_table}{- optional. Provide a data frame with updated KEGG module database OR with custom-made modules.
Default (\code{NULL}; inbuilt data used). See Details.}

\item{use_ko_reference_table}{- optional. Provide a data frame with updated KEGG ortholog database.
Default (\code{NULL}; inbuilt data used). See Details.}

\item{use_enzyme_reference_table}{- optional. Provide a data frame with updated KEGG enzyme database.
Default (\code{NULL}; inbuilt data used). See Details.}
}
\value{
Data frame containing a binary indicator for \code{genes} (rows, specified by rownames) involved in modules (columns, Module IDs specified in names).
A column called '\code{no_match}' is returned if one or more of the genes is not involved in any modules.

\code{NULL} is returned when there are no valid entries to evaluate. Note that enzyme entries are checked for 4 position completeness.
}
\description{
The genes can be either enzymes, given by its Enzyme Classification (EC) number (e.g. "1.1.1.1"),
or KEGG ortholog identifiers (K number, e.g. "K00001").
Using EC or K numbers might give different results, as an EC number might map to multiple K numbers or none.
}
\details{
The \code{use_} set of arguments allow users with KEGG FTP access to provide the updated data from the KEGG databases in the form of reference tables AND/OR for advanced users to provide custom-made modules (see below).
These reference tables can be generated with the \code{parseKEGG} family of functions and need to have a specific format (see function descriptions for details on format).

The module definition (contained in \code{module_reference_table}) describes the relationship between genes and modules
and is used to identify the modules in which \code{genes} is involved.
The user can provide custom-made module definitions that use the logical expression format (however, the table format must be conserved!).

If enzyme identifiers are provided, note that as there are lingering unspecified enzymes (e.g. '1.1.1.-') in the KEGG module EC-based definition, this function also returns a row entry for
the unspecified-versions of enzymes provided involved in one or more modules.
}
\examples{
modules <- query_genes_to_modules("K00844")

modules <- query_genes_to_modules("1.1.1.1")

}
\seealso{
\link{parseKEGG_module}
}

