% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_missingGenes_from_module.R
\name{query_missingGenes_from_module}
\alias{query_missingGenes_from_module}
\title{Identify missing genes from a KEGG module given a set of genes.}
\usage{
query_missingGenes_from_module(GENOME, MODULE_ID, PRINT_TO_SCREEN = TRUE,
  use_genome_reference_table = NULL, use_module_reference_table = NULL, ...)
}
\arguments{
\item{GENOME}{- character vector containing a single genome identifier or set of genes or enzymes that define a [meta]genome. See Details.}

\item{MODULE_ID}{- KEGG module ID to be analysed (e.g. \code{"M00001"}).}

\item{PRINT_TO_SCREEN}{- logical. Should a print-friendly result be displayed? Default(\code{TRUE}).}

\item{use_genome_reference_table}{- optional. Provide a data frame with updated KEGG genome database.
Default (\code{NULL}; inbuilt data used). See Details.}

\item{use_module_reference_table}{- optional. Provide a data frame with updated KEGG module database OR with custom-made modules.
Default (\code{NULL}; inbuilt data used). See Details.}

\item{...}{- further arguments (currently unsupported)}
}
\value{
Data frame containing the following columns:
\preformatted{
 BLOCK_DEF     - the KEGG module DEFINITION of each block,
                   with missing genes flagged by '*';
 PRESENT       - binary indicator of the automatic evaluation;
 MISSING_GENES - list of missing genes.
}
}
\description{
Identify missing genes from a KEGG module given a set of genes or a genome ID to obtain a complete module.
}
\details{
\code{GENOME} can be a genome identifier (T0 number or a 3 or 4 letter code; e.g. "T00001" or "eco", respectively) OR
a character vector contianing a set of genes (i.e. either EC or K numbers; e.g. "1.1.1.1" or "K00001", respectively).
Examples of the latter can be found in the following objects:
\link{data_example_KOnumbers_vector} or \link{data_example_ECnumbers_vector}.

Organism name NOT supported. Use the KEGG database website (\url{http://www.genome.jp/kegg/catalog/org_list.html}) to determine the genome identifier.

Note that the pipe ('|') in the DEFINITION indicates an OR operation
This means that there are several possibile genes that carry out the same reaction or function and only one is required.

The \code{use_} set of arguments allow users with KEGG FTP access to provide the updated data from the KEGG databases in the form of reference tables AND/OR for advanced users to provide custom-made modules (see below).
These reference tables can be generated with the \code{parseKEGG} family of functions and need to have a specific format (see function descriptions for details on format).

The module definition (contained in \code{module_reference_table}) describes the relationship between genes and modules
and is used to identify the modules in which \code{gene} is involved.
The user can provide custom-made module definitions that use the logical expression format (however, the table format must be conserved!).
}
\examples{
# Load data
data("data_example_KOnumbers_vector")
OUT <- query_missingGenes_from_module(data_example_KOnumbers_vector,"M00001")

}
