% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_genomes_module_output.R
\name{analysis_genomes_module_output}
\alias{analysis_genomes_module_output}
\title{Process the output generated by \code{query_genomes_to_modules()}.}
\usage{
analysis_genomes_module_output(FRACTION_MATRIX, QUERIES = NULL,
  outPath = "", report_file = "report.tex", figType = c(".eps", ".png"),
  FACTOR = NULL, ...)
}
\arguments{
\item{FRACTION_MATRIX}{- matrix. $MATRIX matrix generated by \code{query_genomes_to_modules()}.}

\item{QUERIES}{- optional. $QUERIES data frame generated by \code{query_genomes_to_modules()} containing the search query performed.}

\item{outPath}{- optional. String to indicate path to store output file and figures. Default (\code{Sys.Date()} as 'YYYY_MM_DD', e.g. '2000_01_31').}

\item{report_file}{- optional. File name for LaTex report (see \code{Details} for path handling). Default ("report.tex" saved to outPAth; use \code{NULL} to supress report generation).}

\item{figType}{- optional. Character vector to indicate file extension for figures. Default (c(".eps", ".png")).}

\item{FACTOR}{- optional. Character vector or list of character vectors indicating the grouping factor of the datasets. Default (\code{NULL}).See \code{Details}.}

\item{...}{- further arguments for \link{analysis_pca_mean_distance_grouping} or argument { }{ }\code{factor_labs} for \link{plot_scatter_byFactors}.}
}
\value{
This function does not return anything, but does generate multiple files (see \code{Details}).
This function automatically generates a ".tex" file (LaTex report), which can be externally processed to generate a PDF file, unless \code{report_file} is set to \code{NULL}.

ggplot objects for all figures are generated and saved in an R object ("module_output_plots.rda") within the output directory.
}
\description{
Process the output generated by \code{query_genomes_to_modules()}.
}
\details{
Below are the steps carried on the \code{query_genomes_to_modules()} output.

(1) Module fraction completeness (mfc) visualization of the \code{query_genomes_to_modules()} output; generates file: module_allOrgs\code{.figType}.

(2) variance analysis to identify complete and absent modules,
as well as modules with zero variance (i.e. same fraction present across all modules);
generates files: module_allOrgs_sd\code{.figType}, module_allOrgs_sd_boxplot\code{.figType} and module_constant_presence.txt.

(3) a principal component analysis (PCA; using the \code{prcomp} from the 'stats' package) is done
and a plot of the cumulative variance and the first to principal components are generated;
generates files in PCA/ : pca_plot\code{.figType}, pca_sd\code{.figType} and pca.rda

When \code{FACTOR} is specified, a mean distance is calculated as a proxy for inner-group spread
(using as many dimensions as specified by the optional additional argument \code{nDim}; used by \link{analysis_pca_mean_distance_grouping()}).
This generates additional files: module_mean_dist_output_FACTOR.rda and the following for each factor (multiple factors possible if \code{FACTOR} is a list):

module_<FACTOR>_mean.png,module_<FACTOR>_sd\code{.figType},

PCA/plot_mean_dist_<FACTOR>\code{.figType} and PCA/plot_scatter_<FACTOR>\code{.figType}

It will be assumed that if \code{report_file} has a folder structure (identified by "\\") a path has been given.
Otherwise, the report will be written to \code{outPath}. If \code{report_file} is set to \code{NULL}, the report is not generated.
}
\seealso{
\link{query_genomes_to_modules}, \link{analysis_pca_mean_distance_grouping},\link{plot_heatmap}
}
