% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_pca_mean_distance_calculation.R
\name{analysis_pca_mean_distance_calculation}
\alias{analysis_pca_mean_distance_calculation}
\title{Given a set of coordinates, calculate the mean distance between all points.}
\usage{
analysis_pca_mean_distance_calculation(MATRIX, ...)
}
\arguments{
\item{MATRIX}{- matrix with the rows refering to the points with N columns containing the coordinates (and therefore N dimensions).}

\item{...}{- further arguments (currently unsupported)}
}
\value{
The mean distance (\code{numeric}).
}
\description{
Given a set of coordinates, calculate the mean distance between all points.
}
\details{
The mean distance of \eqn{p} points is calculated by the sum of the individual Euclidean distances
divided by the total number of distances (given by \eqn{p*(p-1)/2}).
}
\examples{
data(data_example_moduleIDs)
data(data_example_genomeIDs)

# Calculate the module completion fraction (mcf) for the genomes
#     and modules contained in the data objects above.
OUT         <- query_genomes_to_modules(data_example_genomeIDs,
                                        MODULE_ID = data_example_moduleIDs)

pca <- prcomp(OUT$MATRIX)

mean_dist <- analysis_pca_mean_distance_calculation(pca$x)
# [1] 0.4805169

}
