% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_ko.R
\name{parseKEGG_ko}
\alias{parseKEGG_ko}
\title{Parse the KEGG orthology (KO) database}
\usage{
parseKEGG_ko(KEGG_path, outDir = "output", verbose = T, ...)
}
\arguments{
\item{KEGG_path}{- string pointing to the location of the KEGG database parent folder. The path to the required file is contained within the function.}

\item{outDir}{- optional. String pointing to the output folder. Default ("output/").}

\item{verbose}{- logical. Should progress be printed to the screen? Default (\code{TRUE}).}

\item{...}{- other arguments for \code{parseKEGG_file()}.}
}
\value{
Generates ko_reference_table (.txt & .rda; saved to \code{'outDir'}) and returns
a data frame with as many rows as entries and the following columns (or variables):
\preformatted{

	    *In all instances, multiple entries in a given column are separated by '[;]'.
}
}
\description{
Read and format the KEGG orthology (KO) database (containing ortholog (gene) information) text file into a reference table.
}
\details{
The columns are automatically generated by the \code{parseKEGG_file} function into variables,
which are further formatted specifically for the KEGG ortholog database.

The text file used is "\code{KEGG_path}/genes/ko/ko".

It decompresses "\code{KEGG_path}/genes/ko.tar.gz" if needed.
}
\examples{
KEGG_path <- "~/KEGG" # MODIFY TO KEGG PARENT FOLDER!
# The parent folder should contain the following (KEGG FTP structure):
#	brite/
#	genes/
#	ligand/
#	medicus/
#	module/
#	pathway/
#	README.kegg
#	RELEASE
#	xml/

ko_reference_table <- parseKEGG_ko(KEGG_path)
 # A .txt file (tab separated) is written to output/ (relative to current working directory)

}
\seealso{
\link{parseKEGG_file}
}
