% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_ko_reaction.R
\name{parseKEGG_ko_reaction}
\alias{parseKEGG_ko_reaction}
\title{Map KEGG orthologs (KOs) to Reaction IDs}
\usage{
parseKEGG_ko_reaction(KEGG_path, outDir = "output", verbose = T)
}
\arguments{
\item{KEGG_path}{- string pointing to the location of the KEGG database parent folder}

\item{outDir}{- string pointing to the output folder. Default ("output/"). \code{NULL} overwrites writting files.}

\item{verbose}{- logical. Should progress be printed to the screen? Default (\code{TRUE}).}
}
\value{
Data frame establishing the relationship between KO numbers and reactions (R number) (binary).
\preformatted{
> ko_reaction_map[1:3,1:3]

              R00005 R00006 R00008
    K00001      0      0      0
    K00002      0      0      0
    K00003      0      0      0
}
}
\description{
Map KEGG orthologs (KOs) to Reaction IDs and format it into a matrix with binary indicator for mapping/relationship.
Generates 'ko_reaction_map' (.txt & .rda).
}
\seealso{
\link{parseKEGG_file.list}
}

