% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseKEGG_process_KEGG_taxonomy.R
\name{parseKEGG_process_KEGG_taxonomy}
\alias{parseKEGG_process_KEGG_taxonomy}
\title{Retrieve the taxonomy listed as part of the KEGG genome database.}
\usage{
parseKEGG_process_KEGG_taxonomy(genome_reference_table,
  taxonomy_header = "TAXONOMY", org_header = 1, ...)
}
\arguments{
\item{genome_reference_table}{- table containing the different genome entries (rows) and data (columns). See \link{parseKEGG_genome}.}

\item{taxonomy_header}{- optional. Character with header name or number indicating column index for taxonomic information. Default ("TAXONOMY").}

\item{org_header}{- optional. Character with header name or number indicating column index for organism name or ID. Use to name the output rownames. Default (1; first column).}

\item{...}{- further arguments (currently unsupported)}
}
\value{
Data frame with 6 columns containing the taxonomic information (columns: KINGDOM, PHYLUM, CLASS, ORDER, FAMILY, GENUS).
The rownames contain the organism information (name or ID) as specified with \code{org_header}.
\code{UNKNOWN} label is added where no information is available.
}
\description{
Retrieve the taxonomy listed as part of the KEGG genome database.
}
\details{
After processing the KEGG genome database, this function can extract the taxonomic information.
This is obtained from splitting the \code{TAXONOMY} column after the "LINEAGE" tag into "words".
This information should specify KINGDOM, PHYLUM, CLASS, ORDER, FAMILY, GENUS (6), but there are ocasionally more or less words than expected.
Therefore NOTE that it is incomplete and inconsistent, most likely because it is derived from multiple sources.
Use with caution.
}
\examples{
# Generate KEGG's genome database table
genome_reference_table <- parseKEGG_genome("~/KEGG")

# Extract taxonomic information from genome_referene_table
TAXONOMY_table <- parseKEGG_process_KEGG_taxonomy(genome_reference_table)
TAXONOMY_table[1,]
#            KINGDOM    PHYLUM             CLASS                  ORDER             FAMILY             GENUS
# T00001     Bacteria    Proteobacteria    Gammaproteobacteria    Pasteurellales    Pasteurellaceae    Haemophilus

}
\seealso{
\link{parseKEGG_genome}
}
