% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(data_in, row_lab = "Genomes", col_lab = "Modules",
  ORDER_MATRIX = FALSE, legend_name = "Module\\ncompleteness\\nfraction\\n",
  set_yLim = FALSE, Filename = "", Width = 24, Height = 18)
}
\arguments{
\item{data_in}{- numeric matrix or 3-column data frame containing the module fraction match of the genomes and modules.}

\item{row_lab}{- optional. String to specify the axis label corresponding to the row values. Default ("Genomes").}

\item{col_lab}{- optional. String to specify the axis label corresponding to the row values. Default ("Modules").}

\item{ORDER_MATRIX}{- logical. Should rows and columns be reorder according to the dendogram (hierarchical clustering). Default (\code{FALSE}).}

\item{legend_name}{- optional. String to specify the legend title. Default ("Fraction\\n matched\\n").}

\item{set_yLim}{- optional. Logical or numeric (length 2) indicating whether to set the y limit of the heatmap scale. Default (\code{FALSE}). See Details.}

\item{Filename}{- optional. A character vector containing the file path to save image(s). The device to save is determined by file extension. Default  ("", i.e. file not written).}

\item{Width}{- optional. Width  size for file. Default (24 in).}

\item{Height}{- optional. Height size for file. Default (18 in).}
}
\value{
ggplot object. Saves figures if Filename is specified.
}
\description{
Plot a heatmap of the the module fraction present across genomes or of the variance across groups.
}
\details{
\code{set_yLim} - #' the default (\code{FALSE}), sets the y limit to \code{c(0, 1)}, while \code{TRUE} to the \code{c(min, max)} of the data.
If numeric, the values given are used for the lower and upper limits respectively.
}

