% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_byFactors.R
\name{plot_scatter_byFactors}
\alias{plot_scatter_byFactors}
\title{Scatter plot with overlapping factors/groups.}
\usage{
plot_scatter_byFactors(MATRIX, FACTOR, factor_labs = NULL, xLab = NULL,
  yLab = NULL, Filename = "plot_scatter_byFactors.pdf", Width = 7,
  Height = 5, ...)
}
\arguments{
\item{MATRIX}{- two column matrix to plot. Only the first two columns will be used.}

\item{FACTOR}{- character vector or list of character vectors used to split the data into groups.}

\item{factor_labs}{- optional. Character vector to distinguish \code{FACTOR} groups. Default (\code{NULL}). See Details.}

\item{Filename}{- optional. Character vector containing the file path to save the image. Must have an extension (see Details).
Default ("plot_scatter_byFactors.pdf"). If set to \code{""}, no file will be written.}

\item{Width}{- Width size for file. Default (7 in).}

\item{Height}{- Height size for file. Default (5 in).}

\item{...}{- further arguments (currently unsupported)}
}
\value{
A list with as many entires as factors is generated (one for each factor) using \code{factor_labs} for the names.
For every \code{FACTOR}, a list will contain:
\preformatted{
 $nGroups - numeric. The number of groups found for that FACTOR
                     NOTE: That entries with "" will be excluded.
 $table   - data frame with the number of entries for each group found.
 $file    - character vector of file name(s).
}
}
\description{
Represent different groups as defined by FACTOR(S) in a scatter plot.
}
\details{
This function is used by \link{analysis_genomes_module_output()} to plot the first two Principal Components (PCs) from the PCA analysis,
overlapping different factors or groups (as determined by \code{FACTOR}).
It is also used by \link{analysis_pca_mean_distance_grouping()} to highlight a single group on PC plot.

\code{factor_labs} is used as an extension for the filename for the plot files.
The names of the \code{FACTOR} object OR "factor" followed by a number
will be used if \code{factor_labs} is not specified (i.e. \code{factor_labs = NULL}).

A plot is only generated for FACTORS with less than 60 groups.
All the data is plotted in grey in the background, with groups being overlayed for each Factor.

 Only the following file types are allowed: pdf, png, svg and jpeg.
}
\examples{
data(data_example_moduleIDs)
data(data_example_genomeIDs)
# length(data_example_genomeIDs) # [1] 25

# Calculate the module completion fraction (mcf) for the genomes
#       and modules contained in the data objects above.
OUT         <- query_genomes_to_modules(data_example_genomeIDs,
                                        MODULE_ID = data_example_moduleIDs)

pca <- prcomp(OUT$MATRIX)

# Make boxplots of the mcf output from query_genomes_to_modules
this_FACTOR <- rep(LETTERS[1:5],length(data_example_genomeIDs)/5)
plot_output <- plot_scatter_byFactors(pca$x[,1:2],FACTOR = this_FACTOR,
                                      factor_labs = "random",
                                      Filename = "plot_scatter_byFactors.png")

# NAs are ommitted, so a single group can be contrasted with overall data
this_FACTOR <- c(rep(NA,20),rep(LETTERS[1],5))
plot_output <- plot_scatter_byFactors(pca$x[,1:2],FACTOR = this_FACTOR,
                                      factor_labs = "group_A",
                                      Filename = "plot_scatter_byFactors_single.png")

}
\seealso{
\link{analysis_genomes_module_output},\link{analysis_pca_mean_distance_grouping}
}
