% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variance_boxplot.R
\name{plot_variance_boxplot}
\alias{plot_variance_boxplot}
\title{Make a boxplot}
\usage{
plot_variance_boxplot(MATRIX_IN, x_lab = "Modules",
  y_lab = "Fraction matched", Filename = "", Width = 24, Height = 18)
}
\arguments{
\item{MATRIX_IN}{- numeric matrix containing the module fraction match of the genomes (rows) and modules (columns).
#The columns are used as factor for the box plot grouping.}

\item{x_lab}{- optional. String to specify the x-axis label corresponding to the row values. Default ("Modules").}

\item{y_lab}{- optional. String to specify the legend title. Default ("Fraction\\n matched").}

\item{Filename}{- optional. A character vector containing the file path to save image(s). The device to save is determined by file extension. Default  ("", i.e. file not written).}

\item{Width}{- Width size for file. Default (24in).}

\item{Height}{- Height size for file. Default (18in).}
}
\value{
ggplot object of plot
}
\description{
Make a boxplot
}

